+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [6.2.4](https://github.com/KongHack/ORM/releases/tag/6.2.4)
+- @GameCharmer Trust Cache option
+- @GameCharmer Updates to example config and documentation
+
+
+
 ## [6.2.3](https://github.com/KongHack/ORM/releases/tag/6.2.3)
 - @GameCharmer Really support backed enums this time